+++ sdk/grpc-gen/gen-milvus/milvus.pb.cc
@@ -1424,7 +1424,7 @@ const ::PROTOBUF_NAMESPACE_ID::uint32 TableStruct_milvus_2eproto::offsets[] PROT
   ~0u,  // no _weak_field_map_
   PROTOBUF_FIELD_OFFSET(::milvus::grpc::HIndexParam, status_),
   PROTOBUF_FIELD_OFFSET(::milvus::grpc::HIndexParam, collection_name_),
-  PROTOBUF_FIELD_OFFSET(::milvus::grpc::HIndexParam, index_type_),
+  PROTOBUF_FIELD_OFFSET(::milvus::grpc::HIndexParam, field_names_),
   PROTOBUF_FIELD_OFFSET(::milvus::grpc::HIndexParam, extra_params_),
 };
 static const ::PROTOBUF_NAMESPACE_ID::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
@@ -1670,93 +1670,95 @@ const char descriptor_table_protodef_milvus_2eproto[] PROTOBUF_SECTION_VARIABLE(
   "GetEntityIDsParam\022\027\n\017collection_name\030\001 \001"
   "(\t\022\024\n\014segment_name\030\002 \001(\t\"=\n\020HDeleteByIDP"
   "aram\022\027\n\017collection_name\030\001 \001(\t\022\020\n\010id_arra"
-  "y\030\002 \003(\003\"\220\001\n\013HIndexParam\022#\n\006status\030\001 \001(\0132"
+  "y\030\002 \003(\003\"\221\001\n\013HIndexParam\022#\n\006status\030\001 \001(\0132"
   "\023.milvus.grpc.Status\022\027\n\017collection_name\030"
-  "\002 \001(\t\022\022\n\nindex_type\030\003 \001(\005\022/\n\014extra_param"
-  "s\030\004 \003(\0132\031.milvus.grpc.KeyValuePair*\206\001\n\010D"
-  "ataType\022\010\n\004NULL\020\000\022\010\n\004INT8\020\001\022\t\n\005INT16\020\002\022\t"
-  "\n\005INT32\020\003\022\t\n\005INT64\020\004\022\n\n\006STRING\020\024\022\010\n\004BOOL"
-  "\020\036\022\t\n\005FLOAT\020(\022\n\n\006DOUBLE\020)\022\n\n\006VECTOR\020d\022\014\n"
-  "\007UNKNOWN\020\217N*C\n\017CompareOperator\022\006\n\002LT\020\000\022\007"
-  "\n\003LTE\020\001\022\006\n\002EQ\020\002\022\006\n\002GT\020\003\022\007\n\003GTE\020\004\022\006\n\002NE\020\005"
-  "*8\n\005Occur\022\013\n\007INVALID\020\000\022\010\n\004MUST\020\001\022\n\n\006SHOU"
-  "LD\020\002\022\014\n\010MUST_NOT\020\0032\350\027\n\rMilvusService\022H\n\020"
-  "CreateCollection\022\035.milvus.grpc.Collectio"
-  "nSchema\032\023.milvus.grpc.Status\"\000\022F\n\rHasCol"
-  "lection\022\033.milvus.grpc.CollectionName\032\026.m"
-  "ilvus.grpc.BoolReply\"\000\022R\n\022DescribeCollec"
-  "tion\022\033.milvus.grpc.CollectionName\032\035.milv"
-  "us.grpc.CollectionSchema\"\000\022Q\n\017CountColle"
-  "ction\022\033.milvus.grpc.CollectionName\032\037.mil"
-  "vus.grpc.CollectionRowCount\"\000\022J\n\017ShowCol"
-  "lections\022\024.milvus.grpc.Command\032\037.milvus."
-  "grpc.CollectionNameList\"\000\022P\n\022ShowCollect"
-  "ionInfo\022\033.milvus.grpc.CollectionName\032\033.m"
-  "ilvus.grpc.CollectionInfo\"\000\022D\n\016DropColle"
-  "ction\022\033.milvus.grpc.CollectionName\032\023.mil"
-  "vus.grpc.Status\"\000\022=\n\013CreateIndex\022\027.milvu"
-  "s.grpc.IndexParam\032\023.milvus.grpc.Status\"\000"
-  "\022G\n\rDescribeIndex\022\033.milvus.grpc.Collecti"
-  "onName\032\027.milvus.grpc.IndexParam\"\000\022\?\n\tDro"
-  "pIndex\022\033.milvus.grpc.CollectionName\032\023.mi"
-  "lvus.grpc.Status\"\000\022E\n\017CreatePartition\022\033."
-  "milvus.grpc.PartitionParam\032\023.milvus.grpc"
-  ".Status\"\000\022E\n\014HasPartition\022\033.milvus.grpc."
-  "PartitionParam\032\026.milvus.grpc.BoolReply\"\000"
-  "\022K\n\016ShowPartitions\022\033.milvus.grpc.Collect"
-  "ionName\032\032.milvus.grpc.PartitionList\"\000\022C\n"
-  "\rDropPartition\022\033.milvus.grpc.PartitionPa"
-  "ram\032\023.milvus.grpc.Status\"\000\022<\n\006Insert\022\030.m"
-  "ilvus.grpc.InsertParam\032\026.milvus.grpc.Vec"
-  "torIds\"\000\022J\n\016GetVectorsByID\022\034.milvus.grpc"
-  ".VectorsIdentity\032\030.milvus.grpc.VectorsDa"
-  "ta\"\000\022H\n\014GetVectorIDs\022\036.milvus.grpc.GetVe"
-  "ctorIDsParam\032\026.milvus.grpc.VectorIds\"\000\022B"
-  "\n\006Search\022\030.milvus.grpc.SearchParam\032\034.mil"
-  "vus.grpc.TopKQueryResult\"\000\022J\n\nSearchByID"
-  "\022\034.milvus.grpc.SearchByIDParam\032\034.milvus."
-  "grpc.TopKQueryResult\"\000\022P\n\rSearchInFiles\022"
-  "\037.milvus.grpc.SearchInFilesParam\032\034.milvu"
-  "s.grpc.TopKQueryResult\"\000\0227\n\003Cmd\022\024.milvus"
-  ".grpc.Command\032\030.milvus.grpc.StringReply\""
-  "\000\022A\n\nDeleteByID\022\034.milvus.grpc.DeleteByID"
-  "Param\032\023.milvus.grpc.Status\"\000\022G\n\021PreloadC"
-  "ollection\022\033.milvus.grpc.CollectionName\032\023"
-  ".milvus.grpc.Status\"\000\022I\n\016ReloadSegments\022"
-  " .milvus.grpc.ReLoadSegmentsParam\032\023.milv"
-  "us.grpc.Status\"\000\0227\n\005Flush\022\027.milvus.grpc."
-  "FlushParam\032\023.milvus.grpc.Status\"\000\022=\n\007Com"
-  "pact\022\033.milvus.grpc.CollectionName\032\023.milv"
-  "us.grpc.Status\"\000\022E\n\026CreateHybridCollecti"
-  "on\022\024.milvus.grpc.Mapping\032\023.milvus.grpc.S"
-  "tatus\"\000\022L\n\023HasHybridCollection\022\033.milvus."
-  "grpc.CollectionName\032\026.milvus.grpc.BoolRe"
-  "ply\"\000\022J\n\024DropHybridCollection\022\033.milvus.g"
-  "rpc.CollectionName\032\023.milvus.grpc.Status\""
-  "\000\022O\n\030DescribeHybridCollection\022\033.milvus.g"
-  "rpc.CollectionName\032\024.milvus.grpc.Mapping"
-  "\"\000\022W\n\025CountHybridCollection\022\033.milvus.grp"
-  "c.CollectionName\032\037.milvus.grpc.Collectio"
-  "nRowCount\"\000\022I\n\025ShowHybridCollections\022\024.m"
-  "ilvus.grpc.Command\032\030.milvus.grpc.Mapping"
-  "List\"\000\022V\n\030ShowHybridCollectionInfo\022\033.mil"
-  "vus.grpc.CollectionName\032\033.milvus.grpc.Co"
-  "llectionInfo\"\000\022M\n\027PreloadHybridCollectio"
-  "n\022\033.milvus.grpc.CollectionName\032\023.milvus."
-  "grpc.Status\"\000\022D\n\014InsertEntity\022\031.milvus.g"
-  "rpc.HInsertParam\032\027.milvus.grpc.HEntityID"
-  "s\"\000\022J\n\016HybridSearchPB\022\033.milvus.grpc.HSea"
-  "rchParamPB\032\031.milvus.grpc.HQueryResult\"\000\022"
-  "F\n\014HybridSearch\022\031.milvus.grpc.HSearchPar"
-  "am\032\031.milvus.grpc.HQueryResult\"\000\022]\n\026Hybri"
-  "dSearchInSegments\022#.milvus.grpc.HSearchI"
-  "nSegmentsParam\032\034.milvus.grpc.TopKQueryRe"
-  "sult\"\000\022E\n\rGetEntityByID\022\034.milvus.grpc.Ve"
-  "ctorsIdentity\032\024.milvus.grpc.HEntity\"\000\022J\n"
-  "\014GetEntityIDs\022\037.milvus.grpc.HGetEntityID"
-  "sParam\032\027.milvus.grpc.HEntityIDs\"\000\022J\n\022Del"
-  "eteEntitiesByID\022\035.milvus.grpc.HDeleteByI"
-  "DParam\032\023.milvus.grpc.Status\"\000b\006proto3"
+  "\002 \001(\t\022\023\n\013field_names\030\003 \003(\t\022/\n\014extra_para"
+  "ms\030\004 \003(\0132\031.milvus.grpc.KeyValuePair*\206\001\n\010"
+  "DataType\022\010\n\004NULL\020\000\022\010\n\004INT8\020\001\022\t\n\005INT16\020\002\022"
+  "\t\n\005INT32\020\003\022\t\n\005INT64\020\004\022\n\n\006STRING\020\024\022\010\n\004BOO"
+  "L\020\036\022\t\n\005FLOAT\020(\022\n\n\006DOUBLE\020)\022\n\n\006VECTOR\020d\022\014"
+  "\n\007UNKNOWN\020\217N*C\n\017CompareOperator\022\006\n\002LT\020\000\022"
+  "\007\n\003LTE\020\001\022\006\n\002EQ\020\002\022\006\n\002GT\020\003\022\007\n\003GTE\020\004\022\006\n\002NE\020"
+  "\005*8\n\005Occur\022\013\n\007INVALID\020\000\022\010\n\004MUST\020\001\022\n\n\006SHO"
+  "ULD\020\002\022\014\n\010MUST_NOT\020\0032\256\030\n\rMilvusService\022H\n"
+  "\020CreateCollection\022\035.milvus.grpc.Collecti"
+  "onSchema\032\023.milvus.grpc.Status\"\000\022F\n\rHasCo"
+  "llection\022\033.milvus.grpc.CollectionName\032\026."
+  "milvus.grpc.BoolReply\"\000\022R\n\022DescribeColle"
+  "ction\022\033.milvus.grpc.CollectionName\032\035.mil"
+  "vus.grpc.CollectionSchema\"\000\022Q\n\017CountColl"
+  "ection\022\033.milvus.grpc.CollectionName\032\037.mi"
+  "lvus.grpc.CollectionRowCount\"\000\022J\n\017ShowCo"
+  "llections\022\024.milvus.grpc.Command\032\037.milvus"
+  ".grpc.CollectionNameList\"\000\022P\n\022ShowCollec"
+  "tionInfo\022\033.milvus.grpc.CollectionName\032\033."
+  "milvus.grpc.CollectionInfo\"\000\022D\n\016DropColl"
+  "ection\022\033.milvus.grpc.CollectionName\032\023.mi"
+  "lvus.grpc.Status\"\000\022=\n\013CreateIndex\022\027.milv"
+  "us.grpc.IndexParam\032\023.milvus.grpc.Status\""
+  "\000\022G\n\rDescribeIndex\022\033.milvus.grpc.Collect"
+  "ionName\032\027.milvus.grpc.IndexParam\"\000\022\?\n\tDr"
+  "opIndex\022\033.milvus.grpc.CollectionName\032\023.m"
+  "ilvus.grpc.Status\"\000\022E\n\017CreatePartition\022\033"
+  ".milvus.grpc.PartitionParam\032\023.milvus.grp"
+  "c.Status\"\000\022E\n\014HasPartition\022\033.milvus.grpc"
+  ".PartitionParam\032\026.milvus.grpc.BoolReply\""
+  "\000\022K\n\016ShowPartitions\022\033.milvus.grpc.Collec"
+  "tionName\032\032.milvus.grpc.PartitionList\"\000\022C"
+  "\n\rDropPartition\022\033.milvus.grpc.PartitionP"
+  "aram\032\023.milvus.grpc.Status\"\000\022<\n\006Insert\022\030."
+  "milvus.grpc.InsertParam\032\026.milvus.grpc.Ve"
+  "ctorIds\"\000\022J\n\016GetVectorsByID\022\034.milvus.grp"
+  "c.VectorsIdentity\032\030.milvus.grpc.VectorsD"
+  "ata\"\000\022H\n\014GetVectorIDs\022\036.milvus.grpc.GetV"
+  "ectorIDsParam\032\026.milvus.grpc.VectorIds\"\000\022"
+  "B\n\006Search\022\030.milvus.grpc.SearchParam\032\034.mi"
+  "lvus.grpc.TopKQueryResult\"\000\022J\n\nSearchByI"
+  "D\022\034.milvus.grpc.SearchByIDParam\032\034.milvus"
+  ".grpc.TopKQueryResult\"\000\022P\n\rSearchInFiles"
+  "\022\037.milvus.grpc.SearchInFilesParam\032\034.milv"
+  "us.grpc.TopKQueryResult\"\000\0227\n\003Cmd\022\024.milvu"
+  "s.grpc.Command\032\030.milvus.grpc.StringReply"
+  "\"\000\022A\n\nDeleteByID\022\034.milvus.grpc.DeleteByI"
+  "DParam\032\023.milvus.grpc.Status\"\000\022G\n\021Preload"
+  "Collection\022\033.milvus.grpc.CollectionName\032"
+  "\023.milvus.grpc.Status\"\000\022I\n\016ReloadSegments"
+  "\022 .milvus.grpc.ReLoadSegmentsParam\032\023.mil"
+  "vus.grpc.Status\"\000\0227\n\005Flush\022\027.milvus.grpc"
+  ".FlushParam\032\023.milvus.grpc.Status\"\000\022=\n\007Co"
+  "mpact\022\033.milvus.grpc.CollectionName\032\023.mil"
+  "vus.grpc.Status\"\000\022E\n\026CreateHybridCollect"
+  "ion\022\024.milvus.grpc.Mapping\032\023.milvus.grpc."
+  "Status\"\000\022L\n\023HasHybridCollection\022\033.milvus"
+  ".grpc.CollectionName\032\026.milvus.grpc.BoolR"
+  "eply\"\000\022J\n\024DropHybridCollection\022\033.milvus."
+  "grpc.CollectionName\032\023.milvus.grpc.Status"
+  "\"\000\022O\n\030DescribeHybridCollection\022\033.milvus."
+  "grpc.CollectionName\032\024.milvus.grpc.Mappin"
+  "g\"\000\022W\n\025CountHybridCollection\022\033.milvus.gr"
+  "pc.CollectionName\032\037.milvus.grpc.Collecti"
+  "onRowCount\"\000\022I\n\025ShowHybridCollections\022\024."
+  "milvus.grpc.Command\032\030.milvus.grpc.Mappin"
+  "gList\"\000\022V\n\030ShowHybridCollectionInfo\022\033.mi"
+  "lvus.grpc.CollectionName\032\033.milvus.grpc.C"
+  "ollectionInfo\"\000\022M\n\027PreloadHybridCollecti"
+  "on\022\033.milvus.grpc.CollectionName\032\023.milvus"
+  ".grpc.Status\"\000\022D\n\021CreateHybridIndex\022\030.mi"
+  "lvus.grpc.HIndexParam\032\023.milvus.grpc.Stat"
+  "us\"\000\022D\n\014InsertEntity\022\031.milvus.grpc.HInse"
+  "rtParam\032\027.milvus.grpc.HEntityIDs\"\000\022J\n\016Hy"
+  "bridSearchPB\022\033.milvus.grpc.HSearchParamP"
+  "B\032\031.milvus.grpc.HQueryResult\"\000\022F\n\014Hybrid"
+  "Search\022\031.milvus.grpc.HSearchParam\032\031.milv"
+  "us.grpc.HQueryResult\"\000\022]\n\026HybridSearchIn"
+  "Segments\022#.milvus.grpc.HSearchInSegments"
+  "Param\032\034.milvus.grpc.TopKQueryResult\"\000\022E\n"
+  "\rGetEntityByID\022\034.milvus.grpc.VectorsIden"
+  "tity\032\024.milvus.grpc.HEntity\"\000\022J\n\014GetEntit"
+  "yIDs\022\037.milvus.grpc.HGetEntityIDsParam\032\027."
+  "milvus.grpc.HEntityIDs\"\000\022J\n\022DeleteEntiti"
+  "esByID\022\035.milvus.grpc.HDeleteByIDParam\032\023."
+  "milvus.grpc.Status\"\000b\006proto3"
   ;
 static const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable*const descriptor_table_milvus_2eproto_deps[1] = {
   &::descriptor_table_status_2eproto,
@@ -1816,7 +1818,7 @@ static ::PROTOBUF_NAMESPACE_ID::internal::SCCInfoBase*const descriptor_table_mil
 static ::PROTOBUF_NAMESPACE_ID::internal::once_flag descriptor_table_milvus_2eproto_once;
 static bool descriptor_table_milvus_2eproto_initialized = false;
 const ::PROTOBUF_NAMESPACE_ID::internal::DescriptorTable descriptor_table_milvus_2eproto = {
-  &descriptor_table_milvus_2eproto_initialized, descriptor_table_protodef_milvus_2eproto, "milvus.proto", 8837,
+  &descriptor_table_milvus_2eproto_initialized, descriptor_table_protodef_milvus_2eproto, "milvus.proto", 8908,
   &descriptor_table_milvus_2eproto_once, descriptor_table_milvus_2eproto_sccs, descriptor_table_milvus_2eproto_deps, 50, 1,
   schemas, file_default_instances, TableStruct_milvus_2eproto::offsets,
   file_level_metadata_milvus_2eproto, 51, file_level_enum_descriptors_milvus_2eproto, file_level_service_descriptors_milvus_2eproto,
@@ -21085,6 +21087,7 @@ HIndexParam::HIndexParam()
 HIndexParam::HIndexParam(const HIndexParam& from)
   : ::PROTOBUF_NAMESPACE_ID::Message(),
       _internal_metadata_(nullptr),
+      field_names_(from.field_names_),
       extra_params_(from.extra_params_) {
   _internal_metadata_.MergeFrom(from._internal_metadata_);
   collection_name_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
@@ -21096,16 +21099,13 @@ HIndexParam::HIndexParam(const HIndexParam& from)
   } else {
     status_ = nullptr;
   }
-  index_type_ = from.index_type_;
   // @@protoc_insertion_point(copy_constructor:milvus.grpc.HIndexParam)
 }
 
 void HIndexParam::SharedCtor() {
   ::PROTOBUF_NAMESPACE_ID::internal::InitSCC(&scc_info_HIndexParam_milvus_2eproto.base);
   collection_name_.UnsafeSetDefault(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
-  ::memset(&status_, 0, static_cast<size_t>(
-      reinterpret_cast<char*>(&index_type_) -
-      reinterpret_cast<char*>(&status_)) + sizeof(index_type_));
+  status_ = nullptr;
 }
 
 HIndexParam::~HIndexParam() {
@@ -21133,13 +21133,13 @@ void HIndexParam::Clear() {
   // Prevent compiler warnings about cached_has_bits being unused
   (void) cached_has_bits;
 
+  field_names_.Clear();
   extra_params_.Clear();
   collection_name_.ClearToEmptyNoArena(&::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited());
   if (GetArenaNoVirtual() == nullptr && status_ != nullptr) {
     delete status_;
   }
   status_ = nullptr;
-  index_type_ = 0;
   _internal_metadata_.Clear();
 }
 
@@ -21165,11 +21165,16 @@ const char* HIndexParam::_InternalParse(const char* ptr, ::PROTOBUF_NAMESPACE_ID
           CHK_(ptr);
         } else goto handle_unusual;
         continue;
-      // int32 index_type = 3;
+      // repeated string field_names = 3;
       case 3:
-        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 24)) {
-          index_type_ = ::PROTOBUF_NAMESPACE_ID::internal::ReadVarint(&ptr);
-          CHK_(ptr);
+        if (PROTOBUF_PREDICT_TRUE(static_cast<::PROTOBUF_NAMESPACE_ID::uint8>(tag) == 26)) {
+          ptr -= 1;
+          do {
+            ptr += 1;
+            ptr = ::PROTOBUF_NAMESPACE_ID::internal::InlineGreedyStringParserUTF8(add_field_names(), ptr, ctx, "milvus.grpc.HIndexParam.field_names");
+            CHK_(ptr);
+            if (!ctx->DataAvailable(ptr)) break;
+          } while (::PROTOBUF_NAMESPACE_ID::internal::UnalignedLoad<::PROTOBUF_NAMESPACE_ID::uint8>(ptr) == 26);
         } else goto handle_unusual;
         continue;
       // repeated .milvus.grpc.KeyValuePair extra_params = 4;
@@ -21240,13 +21245,16 @@ bool HIndexParam::MergePartialFromCodedStream(
         break;
       }
 
-      // int32 index_type = 3;
+      // repeated string field_names = 3;
       case 3: {
-        if (static_cast< ::PROTOBUF_NAMESPACE_ID::uint8>(tag) == (24 & 0xFF)) {
-
-          DO_((::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::ReadPrimitive<
-                   ::PROTOBUF_NAMESPACE_ID::int32, ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::TYPE_INT32>(
-                 input, &index_type_)));
+        if (static_cast< ::PROTOBUF_NAMESPACE_ID::uint8>(tag) == (26 & 0xFF)) {
+          DO_(::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::ReadString(
+                input, this->add_field_names()));
+          DO_(::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
+            this->field_names(this->field_names_size() - 1).data(),
+            static_cast<int>(this->field_names(this->field_names_size() - 1).length()),
+            ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::PARSE,
+            "milvus.grpc.HIndexParam.field_names"));
         } else {
           goto handle_unusual;
         }
@@ -21307,9 +21315,14 @@ void HIndexParam::SerializeWithCachedSizes(
       2, this->collection_name(), output);
   }
 
-  // int32 index_type = 3;
-  if (this->index_type() != 0) {
-    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32(3, this->index_type(), output);
+  // repeated string field_names = 3;
+  for (int i = 0, n = this->field_names_size(); i < n; i++) {
+    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
+      this->field_names(i).data(), static_cast<int>(this->field_names(i).length()),
+      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
+      "milvus.grpc.HIndexParam.field_names");
+    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteString(
+      3, this->field_names(i), output);
   }
 
   // repeated .milvus.grpc.KeyValuePair extra_params = 4;
@@ -21352,9 +21365,14 @@ void HIndexParam::SerializeWithCachedSizes(
         2, this->collection_name(), target);
   }
 
-  // int32 index_type = 3;
-  if (this->index_type() != 0) {
-    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::WriteInt32ToArray(3, this->index_type(), target);
+  // repeated string field_names = 3;
+  for (int i = 0, n = this->field_names_size(); i < n; i++) {
+    ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::VerifyUtf8String(
+      this->field_names(i).data(), static_cast<int>(this->field_names(i).length()),
+      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::SERIALIZE,
+      "milvus.grpc.HIndexParam.field_names");
+    target = ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::
+      WriteStringToArray(3, this->field_names(i), target);
   }
 
   // repeated .milvus.grpc.KeyValuePair extra_params = 4;
@@ -21386,6 +21404,14 @@ size_t HIndexParam::ByteSizeLong() const {
   // Prevent compiler warnings about cached_has_bits being unused
   (void) cached_has_bits;
 
+  // repeated string field_names = 3;
+  total_size += 1 *
+      ::PROTOBUF_NAMESPACE_ID::internal::FromIntSize(this->field_names_size());
+  for (int i = 0, n = this->field_names_size(); i < n; i++) {
+    total_size += ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::StringSize(
+      this->field_names(i));
+  }
+
   // repeated .milvus.grpc.KeyValuePair extra_params = 4;
   {
     unsigned int count = static_cast<unsigned int>(this->extra_params_size());
@@ -21409,13 +21435,6 @@ size_t HIndexParam::ByteSizeLong() const {
     total_size += 1 +
       ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::MessageSize(
         *status_);
-  }
-
-  // int32 index_type = 3;
-  if (this->index_type() != 0) {
-    total_size += 1 +
-      ::PROTOBUF_NAMESPACE_ID::internal::WireFormatLite::Int32Size(
-        this->index_type());
   }
 
   int cached_size = ::PROTOBUF_NAMESPACE_ID::internal::ToCachedSize(total_size);
@@ -21445,6 +21464,7 @@ void HIndexParam::MergeFrom(const HIndexParam& from) {
   ::PROTOBUF_NAMESPACE_ID::uint32 cached_has_bits = 0;
   (void) cached_has_bits;
 
+  field_names_.MergeFrom(from.field_names_);
   extra_params_.MergeFrom(from.extra_params_);
   if (from.collection_name().size() > 0) {
 
@@ -21452,9 +21472,6 @@ void HIndexParam::MergeFrom(const HIndexParam& from) {
   }
   if (from.has_status()) {
     mutable_status()->::milvus::grpc::Status::MergeFrom(from.status());
-  }
-  if (from.index_type() != 0) {
-    set_index_type(from.index_type());
   }
 }
 
@@ -21479,11 +21496,11 @@ bool HIndexParam::IsInitialized() const {
 void HIndexParam::InternalSwap(HIndexParam* other) {
   using std::swap;
   _internal_metadata_.Swap(&other->_internal_metadata_);
+  field_names_.InternalSwap(CastToBase(&other->field_names_));
   CastToBase(&extra_params_)->InternalSwap(CastToBase(&other->extra_params_));
   collection_name_.Swap(&other->collection_name_, &::PROTOBUF_NAMESPACE_ID::internal::GetEmptyStringAlreadyInited(),
     GetArenaNoVirtual());
   swap(status_, other->status_);
-  swap(index_type_, other->index_type_);
 }
 
 ::PROTOBUF_NAMESPACE_ID::Metadata HIndexParam::GetMetadata() const {