+++ sdk/interface/ConnectionImpl.h
@@ -146,6 +146,9 @@ class ConnectionImpl : public Connection {
     GetHEntityByID(const std::string& collection_name, const std::vector<int64_t>& id_array,
                    HybridQueryResult& result) override;
 
+    Status
+    CreateHybridIndex(const HIndexParam& index_param) override;
+
  private:
     std::shared_ptr<ClientProxy> client_proxy_;
 };